Deprecate TR_X86OpCode
This class can now be replaced by `OMR::X86::InstOpCode`. We simply copy
over all the APIs from the former to the latter. There is no change in
functionality following this commit.

Further we remove OpBinary.cpp and replace it with OMRInstOpCode.cpp and
we get rid of the "_inlines" file for X86Ops.
