+++ compiler/x/codegen/DivideCheckSnippet.cpp
@@ -30,7 +30,6 @@
 #include "ras/Debug.hpp"
 #include "codegen/X86Instruction.hpp"
 #include "codegen/InstOpCode.hpp"
-#include "x/codegen/X86Ops_inlines.hpp"
 #include "env/IO.hpp"
 #include "env/CompilerEnv.hpp"
