+++ compiler/x/codegen/X86BinaryEncoding.cpp
@@ -62,7 +62,6 @@
 #include "runtime/Runtime.hpp"
 #include "x/codegen/X86Instruction.hpp"
 #include "codegen/InstOpCode.hpp"
-#include "x/codegen/X86Ops_inlines.hpp"
 #include "codegen/StaticRelocation.hpp"
 
 #ifdef J9_PROJECT_SPECIFIC
